requirements.txt
streamlit==1.22.0
yt-dlp==2023.3.4
whisper==1.0.0
pinecone-client==2.2.1
openai==0.27.8
pyyaml==6.0
langchain==0.0.184
numpy==1.24.3